b/239107009 Allow settings to be reset to default
If the zone or project uses a non-default setting,
allow the setting to be reset to default at the VM
level.

Introduce a IsSpecified label to distinguish whether
a setting has the default value because (a) it was
explicitly set to default or (b) because it was
not set at all.

Cf #724.