b/372783667 Enable/disable automatic logons (#1512)
Extend the existing `RdpUserAuthenticationBehavior` setting so that it can be used to control automatic logons.

If automatic logons are enabled (the default), IAP Desktop attempts to use save credentials, or will prompt the user to enter or generate new credentials if necessary.

If automatic logons are disabled, IAP Desktop ignores saved credentials and defers all prompting to the RDP control. This behavior ensures compliance with the _Always prompt for password upon connection_ server-side group policy.